Move testing utilities to make them importable from outside
